requirements.txt
torch>=2.0
transformers>=4.0
sentence-transformers>=2.2
scikit-learn>=1.0
pytest>=7.0
PyYAML>=5.4
numpy>=1.19
tqdm>=4.62
matplotlib>=3.5.0